Implemented AbstractSanMoves
